remove tappable; hover only for big screens...
so tappable prevented onClick event bubbling. this caused
the input to not blur(). i do not exactly know if this could be
prevented.
also the felt performance was degraded on a real mobile device.

thats why i removed it again!

also i disabled hovering color for devices with < 900 width. just a
stupid hack for now.